+++ backend/routes/account.routes.js
@@ -6,6 +6,9 @@ const { addressValidator } = require("../models/users/user.validator.js");
 const getUser = require(path.join(__dirname, "../", "controller", "users", "getUser.controller"));
 const { updateUserFields, updateUserPassword } = require("../controller/users/updateUser.contoroller.js");
 const { getAllProducts, getAllSupplierProducts } = require("../controller/products/products.controller.js");
+const deleteUser = require(path.join(__dirname, "../", "controller", "users", "deleteUser.controller"));
+const EditUserById = require(path.join(__dirname, "../", "controller", "users", "EditUserById.controller"));
+const getAllUsers = require(path.join(__dirname, "../", "controller", "users", "getAllUsers.controller"));
 
 router.use("/", (req, res, next) => {
   try {
@@ -42,28 +45,16 @@ router.get("/details", async (req, res, next) => {
   }
 });
 
-router.get("/orders", async(req, res, next) => {
+router.get("/orders", async (req, res, next) => {
   try {
-    let userOrders
+    let userOrders;
     const type = req.type;
     const userId = req.session?.data._id;
-    if(type === "customer")
-        userOrders = await Orders.find({user : userId})
-    else if(type === "admin" )
-        userOrders = await Orders.find()
-    else
-        userOrders = [];      
+    if (type === "customer") userOrders = await Orders.find({ user: userId });
+    else if (type === "admin") userOrders = await Orders.find();
+    else userOrders = [];
     console.log(userOrders);
-    res.render(
-      path.join(
-        __dirname,
-        "..",
-        "views",
-        "account",
-        "accountOrders",
-        "accountOrders"
-      ), { type, orders: userOrders }
-    );
+    res.render(path.join(__dirname, "..", "views", "account", "accountOrders", "accountOrders"), { type, orders: userOrders });
   } catch (error) {
     res.status(500).send(error.message);
   }
@@ -123,13 +114,10 @@ router.get("/products", async (req, res, next) => {
     if (type === "admin") {
       products = await getAllProducts();
     }
-    if (type === "supplier"){
+    if (type === "supplier") {
       products = await getAllSupplierProducts(req.userId);
     }
-    res.render(
-      path.join(
-        __dirname, "..", "views", "account", "accountProducts", "accountProducts"
-      ), { type, products });
+    res.render(path.join(__dirname, "..", "views", "account", "accountProducts", "accountProducts"), { type, products });
   } catch (error) {
     res.status(500).send(error.message);
   }
@@ -140,7 +128,7 @@ router.get("/orders/:orderId", (req, res, next) => {
     const type = req.type;
     const { orderId } = req.params;
     res.render(path.join(__dirname, "..", "views", "account", "accountViewOrder", "accountViewOrder"), { type });
-  } catch (error) { }
+  } catch (error) {}
 });
 
 router.get("/whishlist", (req, res, next) => {
@@ -152,6 +140,39 @@ router.get("/whishlist", (req, res, next) => {
     res.render(path.join(__dirname, "..", "views", "account", "accountWishlist", "accountWishlist"));
   } catch (error) {
     res.status(500).send(error.message);
+  }
+});
+
+router.get("/users", async (req, res, next) => {
+  try {
+    const type = req.type;
+    let users = [];
+    if (!["admin"].includes(type)) {
+      res.redirect("/account/");
+    }
+    users = await getAllUsers();
+    res.render(path.join(__dirname, "..", "views", "account", "accountUsers", "accountUsers"), { type, users });
+  } catch (error) {
+    res.status(500).send(error.message);
+  }
+});
+
+router.delete("/:id", async (req, res) => {
+  try {
+    await deleteUser(req, res);
+  } catch (err) {
+    console.log(err.message);
+    res.status(500).send(err.message);
+  }
+});
+
+router.put("/:id", async (req, res) => {
+  try {
+    console.log(req.body);
+    EditUserById(req, res);
+  } catch (err) {
+    console.log(err.message);
+    res.status(500).send(err.message);
   }
 });
 module.exports = router;
